tweaks to scripts

+++ install.sh
@@ -1,7 +1,10 @@
 #!/usr/bin/env bash
 
 # Install Homebrew for dependency management
-/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+if [[ ! $(which brew -A) ]]
+then
+    /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+fi
 
 brew install wget > /dev/null 2>&1
 
@@ -14,25 +17,35 @@ brew unlink php56 > /dev/null 2>&1
 brew install php70
 
 # Install Composer to /usr/local/bin
+if [[ ! $(which composer -A) ]]
+then
+echo "Installing Composer..."
 /usr/local/bin/php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-/usr/local/bin/php -r "if (hash_file('SHA384', 'composer-setup.php') === '92102166af5abdb03f49ce52a40591073a7b859a86e8ff13338cf7db58a19f7844fbc0bb79b2773bf30791e935dbd938') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
-/usr/local/bin/php composer-setup.php
+/usr/local/bin/php -r "if (hash_file('SHA384', 'composer-setup.php') === '92102166af5abdb03f49ce52a40591073a7b859a86e8ff13338cf7db58a19f7844fbc0bb79b2773bf30791e935dbd938') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;" > /dev/null 2>&1
+/usr/local/bin/php composer-setup.php > /dev/null 2>&1
 /usr/local/bin/php -r "unlink('composer-setup.php');"
 
 mv composer.phar /usr/local/bin/composer
 chmod +x /usr/local/bin/composer
+fi
+
+COMPOSER_PATH=$(which composer)
 
 # Download and unpack the latest Valet release
 rm -rf $HOME/.valet-cli
 mkdir $HOME/.valet-cli
-wget https://github.com/laravel/valet/archive/master.tar.gz -O $HOME/.valet-cli/valet.tar.gz
+
+echo "Downloading Valet..."
+wget https://github.com/laravel/valet/archive/master.tar.gz -O $HOME/.valet-cli/valet.tar.gz > /dev/null 2>&1
 tar xvzf $HOME/.valet-cli/valet.tar.gz -C $HOME/.valet-cli --strip 1 > /dev/null 2>&1
 
+rm /usr/local/bin/valet
 ln -s $HOME/.valet-cli/valet /usr/local/bin/valet
 chmod +x /usr/local/bin/valet
 
 # Install Valet's Composer dependencies
-/usr/local/bin/php /usr/local/bin/composer install -d $HOME/.valet-cli
+echo "Installing Valet's Composer dependencies..."
+/usr/local/bin/php $COMPOSER_PATH install -d $HOME/.valet-cli > /dev/null 2>&1
 
 # Run the Valet installation process
 $HOME/.valet-cli/valet install
